Create README.md
Inspired by the traceable analogies between protein sequences and music notes, we developed algorithms for protein music composition of enhanced musicality by fine-tuned protein-to-music mapping guided by the features of specific music style. We analyzed the musical features of a specific music style (Fantasy-Impromptu style), and used the quantized features to guide a broad exploration of diverse amino acid properties (104 properties, multiple sequence patterns and variations of biological relevance) for a novel protein-to-music algorithm, aimed at enhanced musicality. This algorithm was applied to 18 proteins of various biological functions, which are provided here